feature: Moved the FSMBuilder and FSMConfig back into the FSM file to be able to convert the FSM into a sealed class.
